reintroduce debug_print file
