+++ unit-tests/util/debug_print.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../src/transaction/transaction_parser.h"  // parse_and_process_instruction_outcome_t
+
+void dbg_print_parse_process_instruction_outcome(parse_and_process_instruction_outcome_t *outcome);
